Move the GenerateServiceDescriptors annotation to a new top-level annotations module with common annotations. Enabled the GenerateServiceDescriptors annotation for any module using service descriptors.
Fixes #12
