+++ annotation/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+/// Module containing common annotations used in Hypospray.
+module eu.jbeernink.hypospray.annotation {
+	exports eu.jbeernink.hypospray.annotation.service to eu.jbeernink.hypospray.core, eu.jbeernink.hypospray.discovery.all, eu.jbeernink.hypospray.discovery.annotated, eu.jbeernink.hypospray.discovery.none, eu.jbeernink.hypospray.model, eu.jbeernink.hypospray.scope.application, eu.jbeernink.hypospray.scope.request;
+}
