--- compile/annotation/src/main/java/module-info.java
@@ -1,3 +1,0 @@
-module eu.jbeernink.hypospray.compile.annotation {
-	exports eu.jbeernink.hypospray.compile.annotation;
-}
