+++ compile/servicedescriptor/src/main/java/module-info.java
@@ -4,6 +4,6 @@ import eu.jbeernink.hypospray.compile.servicedescriptor.ServiceDescriptorGenerat
 
 module eu.jbeernink.hypospray.compile.servicedescriptor {
 	requires java.compiler;
-	requires eu.jbeernink.hypospray.compile.annotation;
+
 	provides Processor with ServiceDescriptorGenerator;
 }
