+++ discovery/annotated/pom.xml
@@ -16,6 +16,12 @@
 	<dependencies>
 		<dependency>
 			<groupId>eu.jbeernink.hypospray</groupId>
+			<artifactId>eu.jbeernink.hypospray.annotation</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>eu.jbeernink.hypospray</groupId>
 			<artifactId>eu.jbeernink.hypospray.core</artifactId>
 			<version>${project.version}</version>
 		</dependency>
@@ -34,6 +40,13 @@
 		<dependency>
 			<groupId>org.jspecify</groupId>
 			<artifactId>jspecify</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>eu.jbeernink.hypospray.compile</groupId>
+			<artifactId>eu.jbeernink.hypospray.compile.servicedescriptor</artifactId>
+			<version>${project.version}</version>
+			<scope>compile</scope>
 		</dependency>
 
 		<dependency>
@@ -87,6 +100,24 @@
 
 	<build>
 		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<configuration>
+					<annotationProcessorPaths>
+						<path>
+							<groupId>eu.jbeernink.hypospray.compile</groupId>
+							<artifactId>eu.jbeernink.hypospray.compile.servicedescriptor</artifactId>
+							<version>${project.version}</version>
+						</path>
+					</annotationProcessorPaths>
+					<annotationProcessors>
+						<annotationProcessor>eu.jbeernink.hypospray.compile.servicedescriptor.ServiceDescriptorGenerator
+						</annotationProcessor>
+					</annotationProcessors>
+				</configuration>
+			</plugin>
+
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
