+++ discovery/none/pom.xml
@@ -15,9 +15,22 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>eu.jbeernink.hypospray</groupId>
+			<artifactId>eu.jbeernink.hypospray.annotation</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
 			<groupId>eu.jbeernink.hypospray.discovery</groupId>
 			<artifactId>eu.jbeernink.hypospray.discovery.scanner</artifactId>
 			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>eu.jbeernink.hypospray.compile</groupId>
+			<artifactId>eu.jbeernink.hypospray.compile.servicedescriptor</artifactId>
+			<version>${project.version}</version>
+			<scope>compile</scope>
 		</dependency>
 
 		<dependency>
@@ -33,4 +46,25 @@
 		</dependency>
 	</dependencies>
 
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<configuration>
+					<annotationProcessorPaths>
+						<path>
+							<groupId>eu.jbeernink.hypospray.compile</groupId>
+							<artifactId>eu.jbeernink.hypospray.compile.servicedescriptor</artifactId>
+							<version>${project.version}</version>
+						</path>
+					</annotationProcessorPaths>
+					<annotationProcessors>
+						<annotationProcessor>eu.jbeernink.hypospray.compile.servicedescriptor.ServiceDescriptorGenerator
+						</annotationProcessor>
+					</annotationProcessors>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
 </project>
