+++ pom.xml
@@ -9,7 +9,9 @@
 	<version>${revision}</version>
 	<packaging>pom</packaging>
 	<modules>
+		<module>annotation</module>
 		<module>codegeneration</module>
+		<module>compile</module>
 		<module>core</module>
 		<module>demo</module>
 		<module>discovery</module>
@@ -23,8 +25,7 @@
 		<module>event</module>
 		<module>model</module>
 		<module>invoker</module>
-        <module>compile</module>
-    </modules>
+	</modules>
 
 	<distributionManagement>
 		<repository>
@@ -135,37 +136,38 @@
 						<consoleOutputReporter>
 							<disable>true</disable>
 						</consoleOutputReporter>
-						<statelessTestsetInfoReporter implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5StatelessTestsetInfoTreeReporter" />
+						<statelessTestsetInfoReporter
+								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5StatelessTestsetInfoTreeReporter"/>
 						<statelessTestsetReporter
-							implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5Xml30StatelessReporter">
+								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5Xml30StatelessReporter">
 							<version>3.0.2</version>
 							<usePhrasedFileName>false</usePhrasedFileName>
 							<usePhrasedTestSuiteClassName>true</usePhrasedTestSuiteClassName>
 							<usePhrasedTestCaseClassName>true</usePhrasedTestCaseClassName>
 							<usePhrasedTestCaseMethodName>true</usePhrasedTestCaseMethodName>
 						</statelessTestsetReporter>
-<!--						<statelessTestsetReporter-->
-<!--								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5Xml30StatelessReporter">-->
-<!--							<disable>false</disable>-->
-<!--							<version>3.0</version>-->
-<!--							<usePhrasedFileName>false</usePhrasedFileName>-->
-<!--							<usePhrasedTestSuiteClassName>true</usePhrasedTestSuiteClassName>-->
-<!--							<usePhrasedTestCaseClassName>true</usePhrasedTestCaseClassName>-->
-<!--							<usePhrasedTestCaseMethodName>true</usePhrasedTestCaseMethodName>-->
-<!--						</statelessTestsetReporter>-->
-<!--						<consoleOutputReporter-->
-<!--								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5ConsoleOutputReporter">-->
-<!--							<disable>false</disable>-->
-<!--							<encoding>UTF-8</encoding>-->
-<!--							<usePhrasedFileName>false</usePhrasedFileName>-->
-<!--						</consoleOutputReporter>-->
-<!--						<statelessTestsetInfoReporter-->
-<!--								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5StatelessTestsetInfoReporter">-->
-<!--							<disable>false</disable>-->
-<!--							<usePhrasedFileName>false</usePhrasedFileName>-->
-<!--							<usePhrasedClassNameInRunning>true</usePhrasedClassNameInRunning>-->
-<!--							<usePhrasedClassNameInTestCaseSummary>true</usePhrasedClassNameInTestCaseSummary>-->
-<!--						</statelessTestsetInfoReporter>-->
+						<!--						<statelessTestsetReporter-->
+						<!--								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5Xml30StatelessReporter">-->
+						<!--							<disable>false</disable>-->
+						<!--							<version>3.0</version>-->
+						<!--							<usePhrasedFileName>false</usePhrasedFileName>-->
+						<!--							<usePhrasedTestSuiteClassName>true</usePhrasedTestSuiteClassName>-->
+						<!--							<usePhrasedTestCaseClassName>true</usePhrasedTestCaseClassName>-->
+						<!--							<usePhrasedTestCaseMethodName>true</usePhrasedTestCaseMethodName>-->
+						<!--						</statelessTestsetReporter>-->
+						<!--						<consoleOutputReporter-->
+						<!--								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5ConsoleOutputReporter">-->
+						<!--							<disable>false</disable>-->
+						<!--							<encoding>UTF-8</encoding>-->
+						<!--							<usePhrasedFileName>false</usePhrasedFileName>-->
+						<!--						</consoleOutputReporter>-->
+						<!--						<statelessTestsetInfoReporter-->
+						<!--								implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5StatelessTestsetInfoReporter">-->
+						<!--							<disable>false</disable>-->
+						<!--							<usePhrasedFileName>false</usePhrasedFileName>-->
+						<!--							<usePhrasedClassNameInRunning>true</usePhrasedClassNameInRunning>-->
+						<!--							<usePhrasedClassNameInTestCaseSummary>true</usePhrasedClassNameInTestCaseSummary>-->
+						<!--						</statelessTestsetInfoReporter>-->
 					</configuration>
 				</plugin>
 			</plugins>
